Money Script
The very basics of a basic money script: Adds a counter to the leaderboard in the game (Will have more than one thing in the leaderboard if something is already there) to give the player some money to spend. To use it with text boxes like in restaurants, just connect the two scripts in some loops or something, like if the player has more than X amount of cash, give them the object and deduct that amount of cash from their inventory.